+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preference_activity/composables/WidgetPreview.kt
@@ -38,7 +38,7 @@ fun WidgetPreview(
     val prefsManager = LocalPreferencesManager.current
     val wallpaperManager = WallpaperManager.getInstance(context)
 
-    val date = HijriDate.todayStr(prefsManager)
+    val date = HijriDate.todayFormatted(prefsManager)
 
     val builtinWallpaper = wallpaperManager.builtInDrawable.toBitmap().asImageBitmap()