+++ app/src/main/kotlin/me/amrbashir/hijriwidget/preference_activity/screens/preferences/DayStartScreen.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.material3.TimePicker
@@ -15,6 +16,8 @@ import androidx.compose.material3.rememberTimePickerState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.fromHtml
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -51,6 +54,14 @@ internal fun DayStartScreen() {
                 .padding(16.dp)
         ) {
             TimePicker(state = timePickerState)
+
+            Text(AnnotatedString.fromHtml("""
+                    <p>- AM: Hijri Day starts <b>after</b> Gregorian Day</p>
+                    <p>- PM: Hijri Day starts <b>before</b> Gregorian Day</p>
+                """.trimIndent()),
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
 
             Row(
                 modifier = Modifier.fillMaxWidth(),